pyproject.toml
[project]
name = "spotify-bpm"
version = "0.1.0"
description = "CLI tool to get BPM for each track in a Spotify playlist"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "spotipy>=2.25.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
spotify-bpm = "spotify_bpm:main"
